# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hausa Live Speech-to-Text API
  description: |
    A comprehensive API for Hausa speech-to-text transcription with subscription tiers, 
    points system, and premium features. Supports both file upload and live streaming transcription.
  version: 0.1.0
  contact:
    name: Hausa Speech API Support
    email: support@hausaspeech.com
servers:
  - url: http://localhost:4000
    description: Development server
  - url: https://api.hausaspeech.com
    description: Production server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/auth/register:
    post:
      summary: Register new user
      description: |
        Register a new user with email and password. Creates user profile in MongoDB
        with default free tier subscription. Password is hashed using bcrypt.
      operationId: registerUser
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                password:
                  type: string
                  format: password
                  minLength: 8
                  description: User's password (minimum 8 characters)
                name:
                  type: string
                  description: User's display name
                preferences:
                  $ref: '#/components/schemas/UserPreferences'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/AuthUser'
                  token:
                    type: string
                    description: JWT token for authentication
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/auth/login:
    post:
      summary: Login user
      description: |
        Authenticate user with email and password. Returns user profile data and JWT token.
      operationId: loginUser
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                password:
                  type: string
                  format: password
                  description: User's password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/AuthUser'
                  token:
                    type: string
                    description: JWT token for authentication
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/auth/refresh:
    post:
      summary: Refresh user token
      description: |
        Refresh JWT token and get updated user profile data.
        Requires valid JWT token in Authorization header.
      operationId: refreshToken
      tags: [Authentication]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties: {}
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/AuthUser'
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/auth/profile:
    get:
      summary: Get user profile
      description: Get current authenticated user's profile information
      operationId: getUserProfile
      tags: [Authentication]
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/AuthUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update user profile
      description: Update current user's profile information
      operationId: updateUserProfile
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: User's display name
                preferences:
                  $ref: '#/components/schemas/UserPreferences'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/AuthUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/auth/usage:
    get:
      summary: Get user usage statistics
      description: Get current user's usage statistics and limits
      operationId: getUserUsage
      tags: [Authentication]
      responses:
        '200':
          description: Usage statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  stats:
                    $ref: '#/components/schemas/UsageStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/auth/subscription/upgrade:
    post:
      summary: Upgrade subscription
      description: Upgrade user to a higher subscription tier
      operationId: upgradeSubscription
      tags: [Subscription]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tier]
              properties:
                tier:
                  type: string
                  enum: [basic, gold, premium]
                  description: Target subscription tier
                paymentMethod:
                  $ref: '#/components/schemas/PaymentMethod'
      responses:
        '200':
          description: Subscription upgraded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/AuthUser'
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/subscription/cancel:
    post:
      summary: Cancel subscription
      description: Cancel current user's subscription
      operationId: cancelSubscription
      tags: [Subscription]
      responses:
        '200':
          description: Subscription cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Points System Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token for MongoDB authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for service-to-service authentication
  schemas:
    AuthUser:
      type: object
      properties:
        id:
          type: string
          description: Unique user ID
        email:
          type: string
          format: email
          description: User's email address
        name:
          type: string
          description: User's display name
        subscriptionTier:
          type: string
          enum: [free, basic, gold, premium]
          description: Current subscription tier
        subscriptionStatus:
          type: string
          enum: [active, cancelled, expired, trial]
          description: Subscription status
        pointsBalance:
          type: integer
          description: Current points balance
        isPremium:
          type: boolean
          description: Whether user has premium features
        usageStats:
          $ref: '#/components/schemas/UsageStats'
    UsageStats:
      type: object
      properties:
        dailyMinutes:
          type: integer
          description: Minutes used today
        monthlyMinutes:
          type: integer
          description: Minutes used this month
        totalMinutes:
          type: integer
          description: Total minutes used
        transcriptsCount:
          type: integer
          description: Total transcripts created
    UserPreferences:
      type: object
      properties:
        language:
          type: string
          enum: [ha-NG, en-US, fr-FR, ar]
          description: Preferred language
        theme:
          type: string
          enum: [light, dark]
          description: UI theme preference
        autoPunctuation:
          type: boolean
          description: Enable automatic punctuation
        cloudSync:
          type: boolean
          description: Enable cloud synchronization
...
    PaymentMethod:
      type: object
      properties:
        type:
          type: string
          enum: [card, paypal, apple, google]
          description: Payment method type
        last4:
          type: string
          description: Last 4 digits of card
        brand:
          type: string
          description: Card brand
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string
              details:
                type: object
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string
...